moved utils and constants around
